Update twin.spec
contrib/twin.spec is very old and uses obsolete tags, references non exiting files, etc. This is an attempt to modernize it and bring closer to Fedora guidelines. Since version twin-0.9.0.tar.gz is also problematic for rpmbuild, this .spec clones git, makes .tar.gz. and successfully builds installable RPM package. libltdl comes in many shapes and forms, I tried to fix that with condition that should work at least on Alt, Debian and Fedora. This needs to be polished more, but I hope that is a good starting point – more people will be interested to try twin if they can easily install and deinstall it as a package.